refactor: 누락된 API 문자열 분리
