--- src/entities/user/api/index.ts
@@ -1,1 +1,0 @@
-//
