+++ src/features/comments/api/add-comment.ts
@@ -1,7 +1,8 @@
 import { NewComment } from "@/entities/comment/models/comment.types";
+import { API_URL } from "@/shared/lib/api-path";
 
 export const addCommentAPI = async (newComment: NewComment) => {
-  const response = await fetch("/api/comments/add", {
+  const response = await fetch(API_URL.COMMENTS_ADD, {
     method: "POST",
     headers: { "Content-Type": "application/json" },
     body: JSON.stringify(newComment),
