+++ src/shared/lib/api-path.ts
@@ -4,6 +4,7 @@ import { User } from "@/entities/user/models/user.types";
 
 // SEGMENTS
 const _API = "api";
+const _ADD = "add";
 const _POSTS = "posts";
 const _COMMENTS = "comments";
 const _USERS = "users";
@@ -17,12 +18,14 @@ export const API_URL = {
   // posts
   POSTS: (limit: number, skip: number) => `/${_API}/${_POSTS}?limit=${limit}&skip=${skip}` as const,
   POSTS_ID: (id: Post["id"]) => `/${_API}/${_POSTS}/${id}` as const,
+  POSTS_ADD: `/${_API}/${_POSTS}/${_ADD}`,
   POSTS_TAG: (tag: Post["tags"][number]) => `/${_API}/${_POSTS}/${_TAG}/${tag}` as const,
   POSTS_TAGS: `/${_API}/${_POSTS}/${_TAGS}`,
   POSTS_SEARCH: (query: string) => `/${_API}/${_POSTS}/${_SEARCH}?q=${query}` as const,
 
   // comments
   COMMENTS_ID: (id: Comment["id"]) => `/${_API}/${_COMMENTS}/${id}` as const,
+  COMMENTS_ADD: `/${_API}/${_COMMENTS}/${_ADD}`,
   COMMENTS_POST_ID: (id: Post["id"]) => `/${_API}/${_COMMENTS}/${_POST}/${id}` as const,
 
   // users
